server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
crewai==0.28.0
crewai-tools==0.1.5
litellm
redis==5.0.1
python-dotenv==1.0.0
httpx==0.25.0
websockets==12.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
aioredis==2.0.1
